refactor(push-notif): export fuction parameters (pushToken, receiver)
