+++ push-notifications/push-notifications.js
@@ -1,0 +1,29 @@
+const rp = require('request-promise');
+
+const sendNotification = async (pushToken, receiver) => {
+  try {
+    const body = {
+      to: pushToken,
+      title: 'Wishopper',
+      body: `Its ${receiver} birthday, why not get her something?`
+    };
+    const options = {
+      method: 'POST',
+      uri: 'https://exp.host/--/api/v2/push/send',
+      headers: {
+        Accept: 'application/json',
+        'Accept-Encoding': 'gzip, deflate',
+        'Content-Type': 'application/json'
+      },
+      body: body,
+      json: true // Automatically stringifies the body to JSON
+    };
+
+    const sendNotification = await rp(options);
+    console.log(`push-notif-status: ${sendNotification.data.status}`); // eslint-disable-line no-console
+  } catch (e) {
+    console.log(e); // eslint-disable-line no-console
+  }
+};
+
+module.exports = sendNotification;
